+++ packages/react/src/components/EmbeddedIFrameWrapper.tsx
@@ -109,7 +109,7 @@ export const EmbeddedIFrameWrapper = (
         />
       )}
       <iframe
-        allow="clipboard-read; clipboard-write"
+        allow="clipboard-read; clipboard-write; sync-xhr 'self' '*.flatfile.com'"
         className={mountElement}
         data-testid={mountElement}
         ref={iframe}