+++ packages/react/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @flatfile/react
 
+## 7.13.7
+
+### Patch Changes
+
+- d145e67: Update sync-xhr settings for embedded iframe
+
+## 7.13.6
+
+### Patch Changes
+
+- 9709f6a: Fix overlap flatfile close button
+
+## 7.13.5
+
+### Patch Changes
+
+- dfe1b5f: Adds externalActorId to the React and Javascript packages to enable repeat guests for portal users
+- Updated dependencies [dfe1b5f]
+  - @flatfile/embedded-utils@1.3.6
+
 ## 7.13.4
 
 ### Patch Changes