+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 1.0.17
 
+## 1.0.22
+
+### Patch Changes
+
+- d145e67: Update sync-xhr settings for embedded iframe
+
 ## 1.0.21
 
 ### Patch Changes